Add ManagedDependenciesPathDetector unit test
